fix sql

+++ script/sql/dinky-pg.sql
@@ -2161,19 +2161,19 @@ COMMENT ON COLUMN "public"."dinky_sys_login_log"."is_deleted" IS 'is deleted';
 DROP TABLE IF EXISTS "public"."dinky_sys_operate_log";
 
 create table  "public"."dinky_sys_operate_log" (
-  id bigint,
+  id int4,
   module_name character varying(50),
-  business_type integer,
+  business_type int4,
   method character varying(100),
   request_method character varying(10),
   operate_name character varying(50),
-  operate_user_id integer,
+  operate_user_id int4,
   operate_url character varying(255),
   operate_ip character varying(50),
   operate_location character varying(255),
   operate_param character varying(2000),
   json_result text,
-  status integer,
+  status int4,
   error_msg text,
   operate_time timestamp without time zone
 );
